Create seperate environment for full test suite run

+++ webapp/features/support/enhanced_support_stop.sh
@@ -1,6 +1,31 @@
-#!/bin/sh
+#!/bin/bash
 
-echo "Killing web server, selenium, Xvfb, Firefox"
-PIDS=`ps aux | grep -i "-e feature -p 8080 -P /trisano\|sel\|Xvfb\|firefox-36/firefox-bin" | grep -v grep | awk '{print $2}'`
-echo $PIDS
-kill $PIDS
+set -e
+
+function kill_pids {
+  if [ "$1" = "" ]
+  then
+    echo "No process found."
+  else
+    echo "pid $1"
+    kill $PIDS
+  fi
+}
+
+
+echo -n "Killing web server on 8080: "
+PIDS=`ps aux | grep -i "-e feature -p 8080 -P /trisano" | grep -v grep | awk '{print $2}'`
+kill_pids $PIDS
+
+echo -n "Killing selenium on 4444: "
+PIDS=`ps aux |grep -i "selenium-server.jar.txt -port 4444" | grep -v grep | awk '{print $2}'`
+kill_pids $PIDS
+
+echo -n "Killing Xvfb :99: "
+PIDS=`ps aux | grep -i "Xvfb :99" | grep -v grep | awk '{print $2}'`
+kill_pids $PIDS
+
+echo -n "Killing Firefox 3.6: "
+PIDS=`ps aux | grep -i "firefox-36/firefox-bin" | grep -v grep | awk '{print $2}'`
+kill_pids $PIDS
+
